// teams/18-Avault/src/avault-shiden-web/src/config/constants/types.ts
export type PageMeta = {
  title: string;
  description?: string;
  image?: string;
};